update Docling and use factories
