+++ docling_serve/datamodel/engines.py
@@ -1,10 +1,4 @@
 import enum
-from typing import Optional
-
-from pydantic import BaseModel
-
-from docling_serve.datamodel.requests import ConvertDocumentsRequest
-from docling_serve.datamodel.responses import ConvertDocumentResponse
 
 
 class TaskStatus(str, enum.Enum):
@@ -18,13 +12,3 @@ class AsyncEngine(str, enum.Enum):
     LOCAL = "local"
 
 
-class Task(BaseModel):
-    task_id: str
-    task_status: TaskStatus = TaskStatus.PENDING
-    request: Optional[ConvertDocumentsRequest]
-    result: Optional[ConvertDocumentResponse] = None
-
-    def is_completed(self) -> bool:
-        if self.task_status in [TaskStatus.SUCCESS, TaskStatus.FAILURE]:
-            return True
-        return False
