+++ docling_serve/settings.py
@@ -31,6 +31,7 @@ class DoclingServeSettings(BaseSettings):
     enable_ui: bool = False
     artifacts_path: Optional[Path] = None
     options_cache_size: int = 2
+    allow_external_plugins: bool = False
 
     cors_origins: list[str] = ["*"]
     cors_methods: list[str] = ["*"]
